requirements.txt
numpy==1.21
scipy
Pillow
cython
matplotlib
scikit-image==0.16.2
#tensorflow==2.6.0
keras==2.3.1
#opencv-python
h5py==2.10.0
imgaug
IPython[all]